refactor(triggers): move trigger logic from flows to triggers module to
