+++ packages/giselle/src/engine/triggers/delete-trigger.ts
